pyproject.toml
[project]
name = "gorest-api-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faker>=40.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]
test = [
    "pytest>=9.0.2",
]
